requirements.txt
arp==1.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
conf==0.4.1
cryptography==38.0.1
idna==3.3
ifcfg==0.23
ipcalc==1.99.0
netaddr==0.8.0
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
python-nmap==0.7.1
requests==2.28.1
scapy==2.4.5
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.12
